[programmers] 240501 assignment
